move separate transformers into `transformers` sub-package
