+++ src/jvm/main/org/jetbrains/kotlinx/lincheck/transformation/transformers/AtomicPrimitiveMethodTransformers.kt
@@ -8,8 +8,13 @@
  * with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-package org.jetbrains.kotlinx.lincheck.transformation
+package org.jetbrains.kotlinx.lincheck.transformation.transformers
 
+import org.jetbrains.kotlinx.lincheck.transformation.*
+import org.jetbrains.kotlinx.lincheck.transformation.ManagedStrategyMethodVisitor
+import org.jetbrains.kotlinx.lincheck.transformation.copyLocals
+import org.jetbrains.kotlinx.lincheck.transformation.invokeStatic
+import org.jetbrains.kotlinx.lincheck.transformation.loadLocals
 import org.objectweb.asm.Type
 import org.objectweb.asm.Type.INT_TYPE
 import org.objectweb.asm.Type.LONG_TYPE
