+++ src/jvm/main/org/jetbrains/kotlinx/lincheck/transformation/transformers/MethodCallTransformer.kt
@@ -8,14 +8,18 @@
  * with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-package org.jetbrains.kotlinx.lincheck.transformation
+package org.jetbrains.kotlinx.lincheck.transformation.transformers
 
+import org.jetbrains.kotlinx.lincheck.transformation.*
 import org.jetbrains.kotlinx.lincheck.transformation.CoroutineInternalCallTracker.isCoroutineInternalClass
+import org.jetbrains.kotlinx.lincheck.transformation.ManagedStrategyMethodVisitor
+import org.jetbrains.kotlinx.lincheck.transformation.invokeIfInTestingCode
+import org.jetbrains.kotlinx.lincheck.transformation.invokeInIgnoredSection
+import org.jetbrains.kotlinx.lincheck.transformation.storeArguments
 import org.objectweb.asm.Opcodes.*
 import org.objectweb.asm.Type
 import org.objectweb.asm.Type.VOID_TYPE
 import org.objectweb.asm.commons.GeneratorAdapter
-import org.objectweb.asm.commons.InstructionAdapter.OBJECT_TYPE
 import sun.nio.ch.lincheck.*
 
 /**
